reorg + cleaning up getEvents and related functions, constants, and helpers
