+++ constants/addresses.ts
@@ -1,15 +1,9 @@
 import { Hex } from 'viem';
 
-export const sepolia = {
-  ERC721_PRESS_FACTORY_PROXY:
-    '0x6975F5CcD3570dBF460451DC6f49CB7ceED8A369' as Hex,
-  ERC721_PRESS_FACTORY: '0x12E5E4bcE816EeFcEfB09e52cB1169b26a81aA62' as Hex,
-  ROLES_WITH_721_GATE_IMMUTABLE_METADATA_NO_FEES:
-    '0x95425eAEEeA3EB6De04C9d922B73967eAB243476' as Hex,
-  CURATION_DATABASE_V1: '0xbfa24Fab9aC17D5482EF375ed1ac54fA8A175B49' as Hex,
-  CURATION_RENDERER_V1: '0xBBEAf40C931cd3b92235A4B25c5025f07fDCA9C0' as Hex,
-};
+export const AP721_DATABASE_V1 = {
+  optimismGoerli: "0xFE16d7a18A1c00e8f07Ca11A1d29e69A69d67d7b" as Hex
+}
 
-export const optimism_goerli = {
- AP721DATABASEV1: '0xFE16d7a18A1c00e8f07Ca11A1d29e69A69d67d7b' as Hex,
+export const CURATION_DATABASE_V1 = {
+  sepolia: "0xbfa24Fab9aC17D5482EF375ed1ac54fA8A175B49" as Hex
 }
