--- constants/events.ts
@@ -1,23 +1,0 @@
-// export const events = {
-//   CREATE_PRESS: 'Create721Press',
-//   DATA_STORED: 'DataStored',
-//   DATA_REMOVED: 'DataRemoved',
-//   DATA_OVERWRITTEN: 'DataOverwritten',
-//   LOGIC_UPDATED: 'LogicUpdated',
-//   PRESS_INITIALIZED: 'PressInitialized',
-//   RENDERER_UPDATED: 'RendererUpdated',
-// };
-
-export const events = {
- DATA_OVERWRITTEN: 'DataOverwritten',
- DATA_REMOVED: 'DataRemoved',
- DATA_STORED: 'DataStored',
- LOGIC_UPDATED: 'LogicUpdated',
- RENDERER_UPDATED: 'RendererUpdated',
- SETUP_AP721: 'SetupAP721',
- 
-
-
-
-
-}
