+++ constants/index.ts
@@ -1,2 +1,1 @@
 export * from './addresses'
-export * from './events'
